SPRT report
Not quite the 20 elo some other people are getting, but I'll take it.
An SPSA will probably optimize this a little further

```
Elo   | 11.86 +- 6.68 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=16MB
LLR   | 3.01 (-2.94, 2.94) [0.00, 5.00]
Games | N: 4982 W: 1529 L: 1359 D: 2094
Penta | [139, 530, 1009, 648, 165]
https://chess.samroelants.com/test/126/
```

+++ simbelmyne/src/uci.rs
@@ -28,6 +28,7 @@ use crate::search::params::ASPIRATION_MIN_DEPTH;
 use crate::search::params::DEFAULT_TT_SIZE;
 use crate::search::params::DELTA_PRUNING_MARGIN;
 use crate::search::params::DOUBLE_EXT_MARGIN;
+use crate::search::params::DOUBLE_EXT_MAX;
 use crate::search::params::FP_BASE;
 use crate::search::params::FP_MARGIN;
 use crate::search::params::FP_THRESHOLD;
@@ -300,7 +301,7 @@ const UCI_OPTIONS: [UciOption; 25] = [
         option_type: OptionType::Spin {
             min: 0,
             max: 20,
-            default: DOUBLE_EXT_MAX as u8,
+            default: DOUBLE_EXT_MAX as i32,
         }
     },
 ];
